Cleanup Run, here's hoping this darn thing actually compiles.

+++ MODSRC/vazkii/tinkerer/client/core/handler/kami/ToolModeHUDHandler.java
@@ -1,15 +1,15 @@
 /**
  * This class was created by <Vazkii>. It's distributed as
  * part of the ThaumicTinkerer Mod.
- * 
+ *
  * ThaumicTinkerer is Open Source and distributed under a
  * Creative Commons Attribution-NonCommercial-ShareAlike 3.0 License
  * (http://creativecommons.org/licenses/by-nc-sa/3.0/deed.en_GB)
- * 
+ *
  * ThaumicTinkerer is a Derivative Work on Thaumcraft 4.
  * Thaumcraft 4 (c) Azanor 2012
  * (http://www.minecraftforum.net/topic/1585216-)
- * 
+ *
  * File Created @ [Dec 29, 2013, 11:38:29 PM (GMT)]
  */
 package vazkii.tinkerer.client.core.handler.kami;
@@ -57,7 +57,7 @@ public class ToolModeHUDHandler {
 				GL11.glPushMatrix();
 				GL11.glEnable(GL11.GL_BLEND);
 				GL11.glBlendFunc(GL11.GL_SRC_ALPHA, GL11.GL_ONE_MINUS_SRC_ALPHA);
-				int color = Color.getHSBColor((float) Math.cos((double) ClientTickHandler.elapsedTicks / 25D), 0.6F, 1F).getRGB();
+				int color = Color.getHSBColor((float) Math.cos(ClientTickHandler.elapsedTicks / 25D), 0.6F, 1F).getRGB();
 				var8.drawStringWithShadow(currentTooltip, tooltipStartX, tooltipStartY, color + (opacity << 24));
 				GL11.glDisable(GL11.GL_BLEND);
 				GL11.glPopMatrix();
@@ -71,7 +71,7 @@ public class ToolModeHUDHandler {
 			tooltipDisplayTicks = 20;
 		}
 	}
-	
+
 	@SideOnly(Side.CLIENT)
 	public static void clientTick() {
 		if(tooltipDisplayTicks > 0)
