+++ MODSRC/vazkii/tinkerer/common/core/handler/kami/SoulHeartHandler.java
@@ -1,15 +1,15 @@
 /**
  * This class was created by <Vazkii>. It's distributed as
  * part of the ThaumicTinkerer Mod.
- * 
+ *
  * ThaumicTinkerer is Open Source and distributed under a
  * Creative Commons Attribution-NonCommercial-ShareAlike 3.0 License
  * (http://creativecommons.org/licenses/by-nc-sa/3.0/deed.en_GB)
- * 
+ *
  * ThaumicTinkerer is a Derivative Work on Thaumcraft 4.
  * Thaumcraft 4 (c) Azanor 2012
  * (http://www.minecraftforum.net/topic/1585216-)
- * 
+ *
  * File Created @ [Dec 29, 2013, 9:33:23 PM (GMT)]
  */
 package vazkii.tinkerer.common.core.handler.kami;
@@ -93,7 +93,7 @@ public class SoulHeartHandler {
 	protected float applyArmorCalculations(EntityLivingBase entity, DamageSource par1DamageSource, float par2) {
 		if (!par1DamageSource.isUnblockable()) {
 			int i = 25 - entity.getTotalArmorValue();
-			float f1 = par2 * (float)i;
+			float f1 = par2 * i;
 			//			this.damageArmor(par2);
 			par2 = f1 / 25.0F;
 		}
@@ -109,7 +109,7 @@ public class SoulHeartHandler {
 		if (entity.isPotionActive(Potion.resistance) && par1DamageSource != DamageSource.outOfWorld) {
 			i = (entity.getActivePotionEffect(Potion.resistance).getAmplifier() + 1) * 5;
 			j = 25 - i;
-			f1 = par2 * (float)j;
+			f1 = par2 * j;
 			par2 = f1 / 25.0F;
 		}
 
@@ -123,7 +123,7 @@ public class SoulHeartHandler {
 
 			if (i > 0 && i <= 20) {
 				j = 25 - i;
-				f1 = par2 * (float)j;
+				f1 = par2 * j;
 				par2 = f1 / 25.0F;
 			}
