+++ MODSRC/vazkii/tinkerer/common/item/kami/tool/ToolHandler.java
@@ -1,15 +1,15 @@
 /**
  * This class was created by <Vazkii>. It's distributed as
  * part of the ThaumicTinkerer Mod.
- * 
+ *
  * ThaumicTinkerer is Open Source and distributed under a
  * Creative Commons Attribution-NonCommercial-ShareAlike 3.0 License
  * (http://creativecommons.org/licenses/by-nc-sa/3.0/deed.en_GB)
- * 
+ *
  * ThaumicTinkerer is a Derivative Work on Thaumcraft 4.
  * Thaumcraft 4 (c) Azanor 2012
  * (http://www.minecraftforum.net/topic/1585216-)
- * 
+ *
  * File Created @ [Dec 29, 2013, 6:01:31 PM (GMT)]
  */
 package vazkii.tinkerer.common.item.kami.tool;
@@ -48,15 +48,15 @@ public final class ToolHandler {
 		int mode = getMode(tool);
 		tool.setItemDamage(getNextMode(mode));
 	}
-	
+
 	public static boolean isRightMaterial(Material material, Material[] materialsListing) {
 		for(Material mat : materialsListing)
 			if(material == mat)
 				return true;
-		
+
 		return false;
 	}
-	
+
 	public static void removeBlocksInIteration(EntityPlayer player, World world, int x, int y, int z, int xs, int ys, int zs, int xe, int ye, int ze, int lockID, Material[] materialsListing, boolean silk, int fortune) {
 		for(int x1 = xs; x1 < xe; x1++)
 			for(int y1 = ys; y1 < ye; y1++)
@@ -73,7 +73,7 @@ public final class ToolHandler {
 
 		if(lockID != -1 && id != lockID)
 			return;
-		
+
 		int meta = world.getBlockMetadata(x, y, z);
 		Material mat = world.getBlockMaterial(x, y, z);
 		Block block = Block.blocksList[id];
@@ -94,7 +94,7 @@ public final class ToolHandler {
 					world.spawnEntityInWorld(new EntityItem(world, bx + 0.5, by + 0.5, bz + 0.5, stack));
 		}
 	}
-	
+
 	public static String getToolModeStr(IAdvancedTool tool, ItemStack stack) {
 		return StatCollector.translateToLocal("ttmisc.mode." + tool.getType() + "." + ToolHandler.getMode(stack));
 	}
@@ -106,11 +106,11 @@ public final class ToolHandler {
 		float f = 1.0F;
 		float f1 = player.prevRotationPitch + (player.rotationPitch - player.prevRotationPitch) * f;
 		float f2 = player.prevRotationYaw + (player.rotationYaw - player.prevRotationYaw) * f;
-		double d0 = player.prevPosX + (player.posX - player.prevPosX) * (double) f;
-		double d1 = player.prevPosY + (player.posY - player.prevPosY) * (double) f;
+		double d0 = player.prevPosX + (player.posX - player.prevPosX) * f;
+		double d1 = player.prevPosY + (player.posY - player.prevPosY) * f;
 		if (!world.isRemote && player instanceof EntityPlayer)
 			d1 += 1.62D;
-		double d2 = player.prevPosZ + (player.posZ - player.prevPosZ) * (double) f;
+		double d2 = player.prevPosZ + (player.posZ - player.prevPosZ) * f;
 		Vec3 vec3 = world.getWorldVec3Pool().getVecFromPool(d0, d1, d2);
 		float f3 = MathHelper.cos(-f2 * 0.017453292F - (float) Math.PI);
 		float f4 = MathHelper.sin(-f2 * 0.017453292F - (float) Math.PI);
@@ -121,7 +121,7 @@ public final class ToolHandler {
 		double d3 = range;
 		if (player instanceof EntityPlayerMP)
 			d3 = ((EntityPlayerMP) player).theItemInWorldManager.getBlockReachDistance();
-		Vec3 vec31 = vec3.addVector((double) f7 * d3, (double) f6 * d3, (double) f8 * d3);
+		Vec3 vec31 = vec3.addVector(f7 * d3, f6 * d3, f8 * d3);
 		return world.rayTraceBlocks_do_do(vec3, vec31, par3, !par3);
 	}
